Added LSTM model using keras library.
Redefined the countVectorize to return a list at the same length as the dictionary with word counter (instead of replacing every word with it's suitable number from the dictionary). This is the same way as done by sklearn.
**The old countVectorize became countDicVectorize**

train_test_vectorization: Shuffled the dataset inside the function instead of in the start of the code.

Note: With the new vectorizer, the success rate is 95%, but takes to long. Need to change the representation to sparse matrix to save resources.